refactor(fabric): Split fabric.py into a core_types module
Deconstructs the monolithic `fabric.py` file into a dedicated `core_types`
package. Each core class (Statement, Rule, Condition, etc.) is now in its own
file.
