+++ src/logic_fabricator/core_types/forking_strategy.py
@@ -1,0 +1,8 @@
+import enum
+
+
+class ForkingStrategy(enum.Enum):
+    COEXIST = "coexist"
+    PRESERVE = "preserve"
+    PRIORITIZE_NEW = "prioritize_new"
+    PRIORITIZE_OLD = "prioritize_old"
