+++ tests/test_belief_system.py
@@ -1,12 +1,12 @@
 import pytest
 from neo4j import GraphDatabase
 
-from logic_fabricator.fabric import (
+from logic_fabricator.core_types import (
     Statement,
     Rule,
     Condition,
-    ContradictionEngine,
     BeliefSystem,
+    ContradictionEngine,
     ContradictionRecord,
     ForkingStrategy,
     Effect,
